Fix a bug which cause an error during the build under windows.

+++ ViSP/src/robot/simulator-robot/vpSimulatorAfma6.cpp
@@ -1707,7 +1707,7 @@ vpSimulatorAfma6::isInJointLimit ()
   }
   
   if (artNumb != 0)
-    std::cout << "\nWarning: Velocity control stopped: axis " << fabs(artNumb) << " on joint limit!" <<std::endl;
+    std::cout << "\nWarning: Velocity control stopped: axis " << fabs((float)artNumb) << " on joint limit!" <<std::endl;
   
   return artNumb;
 }
